Remove ir/chunitachi as an import type.
